+++ js/ngDialog.js
@@ -1,13 +1,24 @@
 /*
  * ngDialog - easy modals and popup windows
  * http://github.com/likeastore/ngDialog
- * (c) 2013 MIT License, https://likeastore.com
+ * (c) 2013-2014 MIT License, https://likeastore.com
  */
 
-(function (window, angular, undefined) {
+(function (root, factory) {
+    if (typeof module !== 'undefined' && module.exports) {
+        // CommonJS
+        module.exports = factory(require('angular'));
+    } else if (typeof define === 'function' && define.amd) {
+        // AMD
+        define(['angular'], factory);
+    } else {
+        // Global Variables
+        factory(root.angular);
+    }
+}(this, function (angular, undefined) {
 	'use strict';
 
-	var module = angular.module('ngDialog', []);
+	var m = angular.module('ngDialog', []);
 
 	var $el = angular.element;
 	var isDef = angular.isDefined;
@@ -16,7 +27,7 @@
 	var animationEndEvent = 'animationend webkitAnimationEnd mozAnimationEnd MSAnimationEnd oanimationend';
 	var forceBodyReload = false;
 
-	module.provider('ngDialog', function () {
+	m.provider('ngDialog', function () {
 		var defaults = this.defaults = {
 			className: 'ngdialog-theme-default',
 			plain: false,
@@ -74,7 +85,7 @@
 					performCloseDialog: function ($dialog, value) {
 						var id = $dialog.attr('id');
 
-						if (typeof window.Hammer !== 'undefined') {
+						if (typeof $window.Hammer !== 'undefined') {
 							var hammerTime = angular.element($dialog).scope().hammerTime;
 							hammerTime.off('tap', closeByDocumentHandler);
 							hammerTime.destroy && hammerTime.destroy();
@@ -285,8 +296,8 @@
 								}
 							};
 
-							if (typeof window.Hammer !== 'undefined') {
-								var hammerTime = scope.hammerTime = window.Hammer($dialog[0]);
+							if (typeof $window.Hammer !== 'undefined') {
+								var hammerTime = scope.hammerTime = $window.Hammer($dialog[0]);
 								hammerTime.on('tap', closeByDocumentHandler);
 							} else {
 								$dialog.bind('click', closeByDocumentHandler);
@@ -403,7 +414,7 @@
 			}];
 	});
 
-	module.directive('ngDialog', ['ngDialog', function (ngDialog) {
+	m.directive('ngDialog', ['ngDialog', function (ngDialog) {
 		return {
 			restrict: 'A',
 			scope : {
@@ -433,5 +444,4 @@
 			}
 		};
 	}]);
-
-})(window, window.angular);
+}));